feature(data-hub): Turn off Airflow internal authentication
Airflow has its own authentication mechanism, but we actually rely on the cluster OAuth proxy to allow elife staff to access it.

This change should instruct airflow to allow all "guest" (i.e. unauthenticated within airflow) visits to act as administrators, and disables internal authentication as a result. All visits still need to pass the OAuth proxy first, so this is OK.