feat: add epp-image-server kustomization for biophysics-colab
This commit introduces a new Kustomization configuration for the epp-image-server in the biophysics-colab namespace. The image is set to update every minute from the master branch of the epp-image-server repository. Other settings include S3Source for static files, a specific S3 bucket, an automation prefix, and a specific AWS IAM role. The base URL is set to the raw data directory in the enhanced-preprints-data repository on GitHub.