Data Hub: EuropePMC LabsLink: Change schedule to hourly
This is to improve evaluated only search delay.
We believe EuropePMC is reading LabsLink in the evening, but it could be any time.
We are now updating Sciety data hourly.
So running it hourly would make sense.
(The CSV import should be finished 40 past the hour)

The LabsLink pipeline itself only takes a few seconds. It does currently build up the whole XML structure in memory.